moved some function from acp server rs to relavent modules
